+++ projects/_99_publication/monthly_tracker_filled_posts/tests/fargate/utils/test_clean_utils.py
@@ -42,11 +42,6 @@ class TestHasColumnDataSinceDate:
         pl_testing.assert_frame_equal(returned_lf, expected_lf)
 
 
-class TestAddCtFilterConsistentService:
-    def test_identifies_locations_that_are_always_care_home_or_always_non_res(self):
-        pass
-
-
 class TestAddCtFilterDispersionFilter:
     def test_identifies_locations_within_ct_posts_dispersion_boundaries(self):
         pass